Add superclass for included and linked structural elements
